itis-3300-CTRL-ALT-ELITE-Recipe-App-main/src/requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
psycopg==3.2.3
psycopg-pool==3.2.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
